Refactor: replace the contextAPI on the checkout and success pages with Zustand and delete the context file
