+++ src/App.tsx
@@ -1,13 +1,10 @@
 import { BrowserRouter } from 'react-router-dom';
 import { Router } from './Router';
-import { CoffeeContextProvider } from './contexts/CoffeeContext';
 
 function App() {
   return (
     <BrowserRouter>
-      <CoffeeContextProvider>
-        <Router />
-      </CoffeeContextProvider>
+      <Router />
     </BrowserRouter>
   );
 }
